remove tracked collection not used anymore
