+++ core/src/main/java/com/orientechnologies/orient/core/db/record/ORecordLazySet.java
@@ -29,11 +29,10 @@ import com.orientechnologies.orient.core.record.ORecord;
 import com.orientechnologies.orient.core.record.ORecordInternal;
 import com.orientechnologies.orient.core.record.impl.ODocument;
 import com.orientechnologies.orient.core.record.impl.ODocumentInternal;
+import com.orientechnologies.orient.core.record.impl.OSimpleMultiValueChangeListener;
 
-import java.util.Collection;
-import java.util.Iterator;
+import java.util.*;
 import java.util.Map.Entry;
-import java.util.Set;
 
 /**
  * Lazy implementation of Set. Can be bound to a source ORecord object to keep track of changes. This avoid to call the makeDirty()
@@ -42,18 +41,261 @@ import java.util.Set;
  *
  * @author Luca Garulli (l.garulli--(at)--orientdb.com)
  */
-public class ORecordLazySet extends ORecordTrackedSet
-    implements Set<OIdentifiable>, ORecordLazyMultiValue, OIdentityChangeListener {
-  protected boolean autoConvertToRecord = true;
+public class ORecordLazySet extends AbstractCollection<OIdentifiable>
+    implements Set<OIdentifiable>, OTrackedMultiValue<OIdentifiable, OIdentifiable>, ORecordElement, ORecordLazyMultiValue,
+    OIdentityChangeListener {
+
+  protected              boolean                    autoConvertToRecord = true;
+  protected final        ORecord                    sourceRecord;
+  protected              Map<OIdentifiable, Object> map                 = new HashMap<OIdentifiable, Object>();
+  private                STATUS                     status              = STATUS.NOT_LOADED;
+  protected static final Object                     ENTRY_REMOVAL       = new Object();
+  private                boolean                    dirty               = false;
+
+  private List<OMultiValueChangeListener<OIdentifiable, OIdentifiable>> changeListeners;
 
   public ORecordLazySet(final ODocument iSourceRecord) {
-    super(iSourceRecord);
+    this.sourceRecord = iSourceRecord;
   }
 
   public ORecordLazySet(ODocument iSourceRecord, Collection<OIdentifiable> iOrigin) {
     this(iSourceRecord);
     if (iOrigin != null && !iOrigin.isEmpty())
       addAll(iOrigin);
+  }
+
+  @Override
+  public ORecordElement getOwner() {
+    return sourceRecord;
+  }
+
+  public boolean addInternal(final OIdentifiable e) {
+    if (map.containsKey(e))
+      return false;
+
+    map.put(e, ENTRY_REMOVAL);
+    addOwnerToEmbeddedDoc(e);
+    return true;
+  }
+
+  @Override
+  public boolean contains(Object o) {
+    return map.containsKey(o);
+  }
+
+  public void clear() {
+    setDirty();
+    map.clear();
+  }
+
+  public boolean removeAll(final Collection<?> c) {
+    boolean changed = false;
+    for (Object item : c) {
+      if (remove(item))
+        changed = true;
+    }
+
+    if (changed)
+      setDirty();
+
+    return changed;
+  }
+
+  public boolean addAll(final Collection<? extends OIdentifiable> c) {
+    if (c == null || c.size() == 0)
+      return false;
+    boolean convert = false;
+    if (c instanceof OAutoConvertToRecord) {
+      convert = ((OAutoConvertToRecord) c).isAutoConvertToRecord();
+      ((OAutoConvertToRecord) c).setAutoConvertToRecord(false);
+    }
+    for (OIdentifiable o : c) {
+      add(o);
+    }
+
+    if (c instanceof OAutoConvertToRecord) {
+      ((OAutoConvertToRecord) c).setAutoConvertToRecord(convert);
+    }
+    return true;
+  }
+
+  public boolean retainAll(final Collection<?> c) {
+    if (c == null || c.size() == 0)
+      return false;
+
+    if (super.retainAll(c)) {
+      return true;
+    }
+    return false;
+  }
+
+  @Override
+  public int size() {
+    return map.size();
+  }
+
+  @SuppressWarnings("unchecked")
+  public ORecordLazySet setDirty() {
+    if (sourceRecord != null) {
+      sourceRecord.setDirty();
+    }
+    this.dirty = true;
+    return this;
+  }
+
+  @Override
+  public void setDirtyNoChanged() {
+    if (sourceRecord != null)
+      sourceRecord.setDirtyNoChanged();
+  }
+
+  public STATUS getInternalStatus() {
+    return status;
+  }
+
+  public void setInternalStatus(final STATUS iStatus) {
+    status = iStatus;
+  }
+
+  public void addChangeListener(final OMultiValueChangeListener<OIdentifiable, OIdentifiable> changeListener) {
+    if (changeListeners == null)
+      changeListeners = new LinkedList<OMultiValueChangeListener<OIdentifiable, OIdentifiable>>();
+    changeListeners.add(changeListener);
+  }
+
+  public void removeRecordChangeListener(final OMultiValueChangeListener<OIdentifiable, OIdentifiable> changeListener) {
+    if (changeListeners != null)
+      changeListeners.remove(changeListener);
+  }
+
+  public Set<OIdentifiable> returnOriginalState(final List<OMultiValueChangeEvent<OIdentifiable, OIdentifiable>> events) {
+    final Set<OIdentifiable> reverted = new HashSet<OIdentifiable>(this);
+
+    final ListIterator<OMultiValueChangeEvent<OIdentifiable, OIdentifiable>> listIterator = events.listIterator(events.size());
+
+    while (listIterator.hasPrevious()) {
+      final OMultiValueChangeEvent<OIdentifiable, OIdentifiable> event = listIterator.previous();
+      switch (event.getChangeType()) {
+      case ADD:
+        reverted.remove(event.getKey());
+        break;
+      case REMOVE:
+        reverted.add(event.getOldValue());
+        break;
+      default:
+        throw new IllegalArgumentException("Invalid change type : " + event.getChangeType());
+      }
+    }
+
+    return reverted;
+  }
+
+  public void fireCollectionChangedEvent(final OMultiValueChangeEvent<OIdentifiable, OIdentifiable> event) {
+    if (changeListeners != null) {
+      for (final OMultiValueChangeListener<OIdentifiable, OIdentifiable> changeListener : changeListeners) {
+        if (changeListener != null)
+          changeListener.onAfterRecordChanged(event);
+      }
+    }
+  }
+
+  protected void addOwnerToEmbeddedDoc(OIdentifiable e) {
+    if (sourceRecord != null && e != null) {
+      ORecordInternal.track(sourceRecord, e);
+    }
+  }
+
+  protected void addEvent(OIdentifiable added) {
+    addOwnerToEmbeddedDoc(added);
+
+    if (changeListeners != null && !changeListeners.isEmpty()) {
+      fireCollectionChangedEvent(
+          new OMultiValueChangeEvent<OIdentifiable, OIdentifiable>(OMultiValueChangeEvent.OChangeType.ADD, added, added));
+    } else {
+      setDirty();
+    }
+  }
+
+  private void updateEvent(OIdentifiable oldValue, OIdentifiable newValue) {
+    if (oldValue instanceof ODocument)
+      ODocumentInternal.removeOwner((ODocument) oldValue, this);
+
+    addOwnerToEmbeddedDoc(newValue);
+
+    if (changeListeners != null && !changeListeners.isEmpty()) {
+      fireCollectionChangedEvent(
+          new OMultiValueChangeEvent<OIdentifiable, OIdentifiable>(OMultiValueChangeEvent.OChangeType.UPDATE, oldValue, newValue,
+              oldValue));
+    } else {
+      setDirty();
+    }
+  }
+
+  protected void removeEvent(OIdentifiable removed) {
+    if (removed instanceof ODocument) {
+      ODocumentInternal.removeOwner((ODocument) removed, this);
+    }
+    if (changeListeners != null && !changeListeners.isEmpty()) {
+      fireCollectionChangedEvent(
+          new OMultiValueChangeEvent<OIdentifiable, OIdentifiable>(OMultiValueChangeEvent.OChangeType.REMOVE, removed, null,
+              removed));
+    } else {
+      setDirty();
+    }
+  }
+
+  @Override
+  public Class<?> getGenericClass() {
+    return OIdentifiable.class;
+  }
+
+  @Override
+  public void replace(OMultiValueChangeEvent<Object, Object> event, Object newValue) {
+    //not needed do nothing
+  }
+
+  private OSimpleMultiValueChangeListener<OIdentifiable, OIdentifiable> changeListener;
+
+  public void enableTracking(ORecordElement parent) {
+    if (changeListener == null) {
+      final OSimpleMultiValueChangeListener<OIdentifiable, OIdentifiable> listener = new OSimpleMultiValueChangeListener<>(this);
+      this.addChangeListener(listener);
+      changeListener = listener;
+      if (this instanceof ORecordLazyMultiValue) {
+        OTrackedMultiValue.nestedEnabled(((ORecordLazyMultiValue) this).rawIterator(), this);
+      } else {
+        OTrackedMultiValue.nestedEnabled(this.iterator(), this);
+      }
+    }
+  }
+
+  public void disableTracking(ORecordElement document) {
+    if (changeListener != null) {
+      final OMultiValueChangeListener<OIdentifiable, OIdentifiable> changeListener = this.changeListener;
+      this.changeListener.timeLine = null;
+      this.changeListener = null;
+      this.dirty = false;
+      removeRecordChangeListener(changeListener);
+      if (this instanceof ORecordLazyMultiValue) {
+        OTrackedMultiValue.nestedDisable(((ORecordLazyMultiValue) this).rawIterator(), this);
+      } else {
+        OTrackedMultiValue.nestedDisable(this.iterator(), this);
+      }
+    }
+  }
+
+  @Override
+  public boolean isModified() {
+    return dirty;
+  }
+
+  @Override
+  public OMultiValueChangeTimeLine<Object, Object> getTimeLine() {
+    if (changeListener == null) {
+      return null;
+    } else {
+      return changeListener.timeLine;
+    }
   }
 
   @Override
@@ -65,7 +307,7 @@ public class ORecordLazySet extends ORecordTrackedSet
   public Iterator<OIdentifiable> iterator() {
     return new OLazyRecordIterator(new OLazyIterator<OIdentifiable>() {
       {
-        iter = ORecordLazySet.super.map.entrySet().iterator();
+        iter = ORecordLazySet.this.map.entrySet().iterator();
       }
 
       private Iterator<Entry<OIdentifiable, Object>> iter;
@@ -119,7 +361,7 @@ public class ORecordLazySet extends ORecordTrackedSet
 
   @Override
   public Iterator<OIdentifiable> rawIterator() {
-    return new OLazyRecordIterator(super.iterator(), false);
+    return new OLazyRecordIterator(new ORecordTrackedIterator(sourceRecord, map.keySet().iterator()), false);
   }
 
   @Override
@@ -237,12 +479,6 @@ public class ORecordLazySet extends ORecordTrackedSet
   @Override
   public int hashCode() {
     return 0;
-  }
-
-  protected void addOwnerToEmbeddedDoc(OIdentifiable e) {
-    if (sourceRecord != null && e != null) {
-      ORecordInternal.track(sourceRecord, e);
-    }
   }
 
 }
